Update GCP and GA configuration and add dataset creation
Updated the Google Cloud Platform and Google Analytics configuration with actual key file path, view ID, project, dataset and table details. Also added a provision for dynamically creating a dataset in BigQuery if it doesn't exist already, improving the usability of the script. This is set to a default location 'EU'.